# backend-tixup-main/backend-tixup-main/src/config/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TixUp API",
    "description": "API para gerenciamento de eventos e ingressos",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "basePath": "",
  "paths": {
...
    "/api/eventos/": {
      "get": {
        "tags": [
          "Eventos"
        ],
        "summary": "Lista todos os eventos públicos",
        "description": "",
        "parameters": [
          {
            "name": "categoria",
            "in": "query",
            "description": "Filtrar por categoria",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "data",
            "in": "query",
            "description": "Filtrar por data (formato YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Eventos públicos listados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Evento"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Eventos"
        ],
        "summary": "Cria um novo evento",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Evento criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evento"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string",
                    "example": "Show do Matuê"
                  },
                  "descricao": {
                    "type": "string",
                    "example": "Evento com atrações especiais"
                  },
                  "data": {
                    "type": "string",
                    "format": "date",
                    "example": "2025-06-20"
                  },
                  "local": {
                    "type": "string",
                    "example": "Ginásio Central"
                  },
                  "preco": {
                    "type": "number",
                    "example": 99.9
                  },
                  "imagem": {
                    "type": "string",
                    "format": "binary"
                  },
                  "publico": {
                    "type": "boolean",
                    "example": true
                  },
                  "categoria": {
                    "type": "string",
                    "example": "show"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/eventos/{id}": {
      "get": {
        "tags": [
          "Eventos"
        ],
        "summary": "Busca um evento específico por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do evento"
          }
        ],
        "responses": {
          "200": {
            "description": "Evento encontrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evento"
                }
              }
            }
          },
          "404": {
            "description": "Evento não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Evento não encontrado"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "NOT_FOUND"
                        },
                        "details": {
                          "type": "string",
                          "example": "Nenhum detalhe adicional disponível"
                        },
                        "suggestion": {
                          "type": "string",
                          "example": "Verifique se o ID do evento está correto."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Eventos"
        ],
        "summary": "Edita um evento existente",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do evento"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evento atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evento"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação nos campos fornecidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Campos obrigatórios: nome, data, local, preco, publico"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "BAD_REQUEST"
                        },
                        "details": {
                          "type": "string",
                          "example": "Campos obrigatórios ausentes"
                        },
                        "suggestion": {
                          "type": "string",
                          "example": "Verifique os campos fornecidos e tente novamente."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Não autorizado a editar este evento",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Você não tem permissão para editar este evento"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "FORBIDDEN"
                        },
                        "details": {
                          "type": "string",
                          "example": "Apenas o organizador do evento pode editá-lo"
                        },
                        "suggestion": {
                          "type": "string",
                          "example": "Verifique suas permissões ou contate o suporte."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Evento não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string",
                    "example": "Show do Matuê - Atualizado"
                  },
                  "descricao": {
                    "type": "string",
                    "example": "Evento com novas atrações"
                  },
                  "data": {
                    "type": "string",
                    "format": "date",
                    "example": "2025-06-21"
                  },
                  "local": {
                    "type": "string",
                    "example": "Estádio Municipal"
                  },
                  "preco": {
                    "type": "number",
                    "example": 120
                  },
                  "imagem": {
                    "type": "string",
                    "example": "https://imagem.com/nova-img.jpg"
                  },
                  "publico": {
                    "type": "boolean",
                    "example": true
                  },
                  "categoria": {
                    "type": "string",
                    "example": "show"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Eventos"
        ],
        "summary": "Deleta um evento por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do evento"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Evento deletado com sucesso"
          },
          "403": {
            "description": "Não autorizado a deletar este evento"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/eventos/meus-eventos": {
      "get": {
        "tags": [
          "Eventos"
        ],
        "summary": "Lista os eventos do organizador logado",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/eventos/{id}/vendas": {
      "get": {
        "tags": [
          "Vendas"
        ],
        "summary": "Lista todos os ingressos vendidos para um evento",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do evento"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de ingressos vendidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "uuid-do-ingresso"
                      },
                      "nome_comprador": {
                        "type": "string",
                        "example": "João Comprador"
                      },
                      "data_compra": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-18T10:00:00Z"
                      },
                      "status": {
                        "type": "string",
                        "example": "pendente"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Apenas organizadores podem visualizar vendas"
          },
          "404": {
            "description": "Evento não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Evento": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "uuid"
                  }
                }
              },
              "nome": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "descricao": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "data": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date"
                  }
                }
              },
              "local": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "preco": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  }
                }
              },
              "imagem": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "publico": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "boolean"
                  }
                }
              },
              "categoria": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "organizador_id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "uuid"
                  }
                }
              },
              "created_at": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  }
                }
              },
              "updated_at": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "Token JWT no formato: Bearer <token>"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}